Add files via upload
Movie correlations generator which uses pandas library to analyze movie ratings and the relationships between how similar users rate the same movies. Generates an excel spreadsheet of movie correlations.. The movie recommender uses the correlations to produce a list of movies that the user will like based on user movie inputs with 94% accuracy.